backend/requirements.txt
fastapi
uvicorn
sqlmodel
google-generativeai
python-dotenv
pydantic
sqlalchemy
